Refactor legendItems import and move to constants for better structure
